strategies moved, producer added to config
