+++ tests/unittests/core/experiment.yaml
@@ -132,6 +132,8 @@
       judgement: ~
       suspend: False
       done: False
+  producer:
+    strategy: NoParallelStrategy
 
 
 - name: supernaedo2.1
@@ -175,7 +177,7 @@
   # user's configuration, but defaults are inferred
   algorithms:
     random: {}
-  
+
 - name: supernaedo2.2
 
   metadata:
@@ -242,7 +244,7 @@
     root_id: supernaedo2
     parent_id: supernaedo2
     adapter:
-      - of_type: dimensionrenaming 
+      - of_type: dimensionrenaming
         old_name: /encoding_layer
         new_name: /encoding
 
@@ -255,7 +257,7 @@
 
   # **complete** specification of algorithms used
   # user's configuration, but defaults are inferred
-  algorithms: 
+  algorithms:
     random: {}
 
 - name: supernaedo2.3.1
@@ -301,7 +303,7 @@
 
   # **complete** specification of algorithms used
   # user's configuration, but defaults are inferred
-  algorithms: 
+  algorithms:
     random: {}
 
 - name: supernaedo2.3.1.1
@@ -340,7 +342,7 @@
 
   # **complete** specification of algorithms used
   # user's configuration, but defaults are inferred
-  algorithms: 
+  algorithms:
     random: {}
 
 - name: supernaedo2.3.1.2
@@ -379,7 +381,7 @@
 
   # **complete** specification of algorithms used
   # user's configuration, but defaults are inferred
-  algorithms: 
+  algorithms:
     random: {}
 
 - name: supernaedo2.3.1.3
@@ -418,7 +420,7 @@
 
   # **complete** specification of algorithms used
   # user's configuration, but defaults are inferred
-  algorithms: 
+  algorithms:
     random: {}
 
 - name: supernaedo2.2.1
